requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
postgres==4.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
psycopg2-pool==1.2
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
